Moves styles out into own folder
