+++ src/components/ControlPanel.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import AddCoinListItem from './AddCoinListItem';
-import './ControlPanel.sass';
+import '../styles/ControlPanel.sass';
 import { Coin } from '../reducers/coins';
 
 interface ControlPanelProps {
