+++ src/styles/Coin.sass
